+++ packages/twenty-sdk/src/cli/utilities/api/api-service.ts
@@ -140,14 +140,8 @@ export class ApiService {
     return this.logicFunctionApi.subscribeToLogs(...args);
   }
 
-  uploadAppTarball(...args: Parameters<FileApi['uploadAppTarball']>) {
-    return this.fileApi.uploadAppTarball(...args);
-  }
-
-  createAppTarballUpload(
-    ...args: Parameters<FileApi['createAppTarballUpload']>
-  ) {
-    return this.fileApi.createAppTarballUpload(...args);
+  createFileUpload(...args: Parameters<FileApi['createFileUpload']>) {
+    return this.fileApi.createFileUpload(...args);
   }
 
   completeAppTarballUpload(