+++ packages/twenty-server/test/integration/metadata/suites/application/utils/create-app-tarball-upload-query-factory.util.ts
@@ -1,13 +1,19 @@
 import gql from 'graphql-tag';
 
 export const createAppTarballUploadQueryFactory = ({
+  filename,
   size,
 }: {
+  filename: string;
   size: number;
 }) => ({
   query: gql`
-    mutation CreateAppTarballUpload($size: Float!) {
-      createAppTarballUpload(size: $size) {
+    mutation CreateAppTarballUpload($filename: String!, $size: Float!) {
+      createFileUpload(
+        filename: $filename
+        size: $size
+        fileFolder: AppTarball
+      ) {
         fileId
         uploadUrl
         contentType
@@ -15,5 +21,5 @@ export const createAppTarballUploadQueryFactory = ({
       }
     }
   `,
-  variables: { size },
+  variables: { filename, size },
 });